GG-34700 Optimized consumption of the memory heap while expiring cache entries.
