+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/GridCacheMapEntry.java
@@ -3892,14 +3892,7 @@ public abstract class GridCacheMapEntry extends GridMetadataAwareAdapter impleme
             if (!(expireTime0 > 0 && expireTime0 <= expireTime))
                 return false;
 
-            CacheObject expiredVal = this.val;
-
-            if (markObsolete0(obsoleteVer, true, null))
-                obsolete = true;
-
-            removeExpiredValue(obsoleteVer);
-
-            if (expiredVal != null) { // Do not trigger events for tombstones.
+            if (this.val != null) { // Do not trigger events for tombstones.
                 if (cctx.events().isRecordable(EVT_CACHE_OBJECT_EXPIRED)) {
                     cctx.events().addEvent(partition(),
                         key,
@@ -3908,16 +3901,21 @@ public abstract class GridCacheMapEntry extends GridMetadataAwareAdapter impleme
                         EVT_CACHE_OBJECT_EXPIRED,
                         null,
                         false,
-                        expiredVal,
-                        expiredVal != null,
+                        this.val,
+                        true,
                         null,
                         null,
                         null,
                         true);
                 }
 
-                cctx.continuousQueries().onEntryExpired(this, key, expiredVal);
+                cctx.continuousQueries().onEntryExpired(this, key, this.val);
             }
+
+            if (markObsolete0(obsoleteVer, true, null))
+                obsolete = true;
+
+            removeExpiredValue(obsoleteVer);
 
             updatePlatformCache(null, null);
         }
